+++ lib/ReactViews/BottomDock/BottomDock.jsx
@@ -6,9 +6,9 @@ import { observer } from "mobx-react";
 import PropTypes from "prop-types";
 import React from "react";
 import ChartPanel from "../Custom/Chart/ChartPanel";
-import MapDataCount from "./MapDataCount";
 import Timeline from "./Timeline/Timeline";
 import Styles from "./bottom-dock.scss";
+import measureElement from "../../ReactViews/HOCs/measureElement";
 
 // import ChartDisclaimer from "./ChartDisclaimer";
 
@@ -19,6 +19,7 @@ const BottomDock = observer(
     propTypes: {
       terria: PropTypes.object.isRequired,
       viewState: PropTypes.object.isRequired,
+      heightFromMeasureElementHOC: PropTypes.number,
       domElementRef: PropTypes.func
     },
 
@@ -26,6 +27,17 @@ const BottomDock = observer(
       runInAction(() => {
         this.props.viewState.topElement = "BottomDock";
       });
+    },
+
+    componentDidUpdate(prevProps) {
+      if (
+        prevProps.heightFromMeasureElementHOC !==
+        this.props.heightFromMeasureElementHOC
+      ) {
+        this.props.viewState.setBottomDockHeight(
+          this.props.heightFromMeasureElementHOC
+        );
+      }
     },
 
     render() {
@@ -39,7 +51,10 @@ const BottomDock = observer(
               ? "top-element"
               : ""
           }`}
-          ref={this.props.domElementRef}
+          ref={element => {
+            this.props.domElementRef(element);
+            this.refToMeasure = element;
+          }}
           tabIndex={0}
           onClick={this.handleClick}
           css={`
@@ -47,7 +62,6 @@ const BottomDock = observer(
           `}
         >
           {/* <ChartDisclaimer terria={terria} viewState={this.props.viewState} /> */}
-          <MapDataCount terria={terria} viewState={this.props.viewState} />
           <ChartPanel
             terria={terria}
             onHeightChange={this.onHeightChange}
@@ -63,4 +77,4 @@ const BottomDock = observer(
   })
 );
 
-module.exports = BottomDock;
+module.exports = measureElement(BottomDock);