+++ lib/ReactViews/Workbench/Controls/ViewingControls.jsx
@@ -202,6 +202,7 @@ const ViewingControls = observer(
         toolName: "Difference",
         getToolComponent: () =>
           import("../../Tools/DiffTool/DiffTool").then(m => m.default),
+        showCloseButton: true,
         params: {
           sourceItem: this.props.item
         }